Remove __init__.py from test directories.
As per py.test's recommendations (and previously brought up in #96).

This is the first step towards making the test suite run against an *installed*
version of the software rather than the copy in which the test code runs. As a
bonus, the test code nolonger gets included in the packaged software.