Separate test stages to work-around py.test path.
In order to cope with pytest's modification of the python module path, the test
suite and doctests must be run separately for system-wide installations of the
library-under-test...

This commit updates tox and travis to do this and notes novelties in the
setup.py.

Additionally, the test-dependencies have been removed from setup.py since there
is no value to them being there.